+++ src/main/java/info/RestaurantInfo.java
@@ -11,7 +11,7 @@ public class RestaurantInfo extends Info implements Comparable<RestaurantInfo> {
 	public String url;
 	
 	
-	public RestaurantInfo(String name, double rating, String placeID, String address, int price, String driveTimeText,
+	public RestaurantInfo(String name, int rating, String placeID, String address, int price, String driveTimeText,
 			int driveTimeValue, String phone, String url) {
 		this.name = name;
 		this.rating = rating;
